Update monkey patches (#2728)
