+++ ultralytics/yolo/utils/ops.py
@@ -1,3 +1,5 @@
+# Ultralytics YOLO 🚀, AGPL-3.0 license
+
 import contextlib
 import math
 import re
